requirements.txt
pillow
watchdog
torch
torchvision
numpy
onnx